+++ date/changelog.md
@@ -1,1 +1,7 @@
 # F1JS Date - Changelog
+
+## 06 Oct 2022 - Ver 1.0.0
+  - Initial version. Created to help with Salon Alure site Calendar
+
+## 13 Nov 2022 - Ver 1.1.0
+  - Add prevDayYmd(), nextDayYmd() and ymd2dmy()